Fix bug in Cost_VAR (p, nr- 1) longest interval
binSeg and PELT assume that intervals are of the form (start, End)
where max(End) = nr and start = 0 - count from 1 cuz 0 is excluded.

However, for cost_var, the longest interval is (p, nr-1) - here counted from 0.

Therefore some adjustments needed to be made, particularly

if(start < p){ //to account for the fact that the longest valid section is y[p:(nr-1)] - here idx from 0
    start = p;
  }

  end = end - 1;
and
  arma::mat Z_sub = Z_full.rows(start - p, end-p);
  arma::mat Y_sub = X.rows(start, end);